Revert build breaking change from Lucene upgrade to 9.3.0

+++ src/main/java/org/opensearch/security/configuration/DlsFlsFilterLeafReader.java
@@ -965,7 +965,7 @@ class DlsFlsFilterLeafReader extends SequentialStoredFieldsLeafReader  {
                     }
 
                     @Override
-                    public int docValueCount() {
+                    public long docValueCount() {
                         return sortedSetDocValues.docValueCount();
                     }
